Read and write *unencrypted* preferences from the snagit prefs directory
Lots going on here. Each command needs to have access to the preferences
as the preferences contain the jira url, username and password. I store
the preferences in a json format and use JSON.NET to serialize and
deserialize the preferences. This works but currently the settings are
stored unencrypted. Do *not* use this anywhere!