Create a new issue and attach an image to it
Works with some caveats. You must specify the project *key*. Not the
project name. You must also specify an issue type that exists and enter
a basic summary. At this time there are no checks to ensure you enter
the required fields or that you have entered them correctly. This will
need to be added. There is also a fair amount of code duplication.