Make Project and IssueType textboxes Listboxes when creating new issues
The available project and issue types are obtained before the new form
is shown directly from Jira. This will ensure that valid selections are
made for these values.

Ultimately, I would like to cache this information so that the app does
not connect to Jira each time a new issue is created. This would speed
up usage.