Rudimentary build script to build the solution via FAKE
Currently this script is very simple. Currently it only cleans the build
directories, stamps the AssemlyInfo.cs file with version info, and
builds the default solution in Release mode. As other functionality is
added to the system this will be expanded to an installer, deploy said
installer, and possibly update and deploy documentation.

Resolves #3