Provide AssemblyInfomationalVersion to contain the git SHA
The intent here is that the git SHA will help identify precisely which
build of the app went into the wild. However, this will not become
easily visible if/until an about box is implemented. The version number
will remain a more user friendly indicator especially in the eventual
release notes.

WIP #1