Only expose SnagitJiraOutputButton and Package to COM
While comparing the outputs of regasm.exe to the outputs of wix heat.exe
I noticed that I was exposing far more classes to COM than were needed.
This change reduces the number of classes that we expose via COM and as
a result reduces the number of registery modifications we need to make.

WIP #1